requirements.txt
flask==3.1.3
pymongo==4.16.0
python-dotenv==1.2.2
flask-cors==6.0.2
